test(task): add e2e tests

+++ src/task/task.service.ts
@@ -1,5 +1,9 @@
-import { Injectable } from '@nestjs/common';
-import { CreateTaskDto } from './dto';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { CreateTaskDto, UpdateTaskDto } from './dto';
 import { PrismaService } from '@/prisma/prisma.service';
 
 @Injectable()
@@ -20,15 +24,50 @@ export class TaskService {
     });
   }
 
-  findOne(id: string) {
-    return `This action returns a #${id} task`;
+  async findOne(id: string, userId: string) {
+    const task = await this.prisma.task.findUnique({
+      where: { id, userId },
+    });
+
+    if (!task) {
+      throw new NotFoundException('Task not found');
+    }
+
+    return task;
   }
 
-  // update(id: string, updateTaskDto: UpdateTaskDto) {
-  //   return `This action updates a #${id} task`;
-  // }
+  async update(id: string, updateTaskDto: UpdateTaskDto, userId: string) {
+    const taskToUpdate = await this.prisma.task.findUnique({
+      where: { id },
+    });
 
-  remove(id: string) {
+    if (!taskToUpdate) {
+      throw new NotFoundException('Task not found');
+    }
+
+    if (taskToUpdate.userId !== userId) {
+      throw new ForbiddenException('You are not allowed to update this task');
+    }
+
+    return this.prisma.task.update({
+      where: { id },
+      data: updateTaskDto,
+    });
+  }
+
+  async remove(id: string, userId: string) {
+    const taskToDelete = await this.prisma.task.findUnique({
+      where: { id },
+    });
+
+    if (!taskToDelete) {
+      throw new NotFoundException('Task not found');
+    }
+
+    if (taskToDelete.userId !== userId) {
+      throw new ForbiddenException('You are not allowed to delete this task');
+    }
+
     return this.prisma.task.delete({
       where: { id },
     });
